requirements.txt
discord.py-self>=2.0.0
rich>=13.0.0